Drop a less important test that is timing-sensitive
GET /api/overview is an endpoint of
secondary (for most users of this library)
importance and a large response which
returns a JSON array (instead of a map) for one
or two metric emission intervals after node
boot.

Statically typed client libraries are sensitive
to such differences, even if they are entirely
transient.

FWIW that endpoint is covered in the rabbitmqadmin
v2 test suite.