Make key delete operations optionally idempotent
In some cases, idempotence is a great property.
For example, queue and exchange deletion operations
have been idempotent in AMQP 0-9-1 since 2013
(RabbitMQ 3.0.0 to be exact).

For example, 'rabbitmqadmin' v2 users might
choose to consider 404 Not Found responses
an error in some cases and a nuisance in others.

Hence rabbitmq/rabbitmqadmin-ng#32.

However, in tests we don't want to ignore 404s
as they can hide an issue in the test itself.