Breaking change: provide more context in error::Error
We'd like top provide the caller with more context
to, say, print key portions of a response in
debug mode or emit metrics.

Unfortunatley, the key reqwest Response methods
consume self and make the value largely
useless for an error that wraps it.

So let's drop the Response part of copy over
the request URL, header map and body. The first
two do not consume self.