Adapt responses::Connection for direct connections
Direct connections in RabbitMQ are
the connections that use the "direct" connectivity
option in the Erlang AMQP 0-9-1 client, that is,
rely on the Erlang distribution protocol over
a net TCP connection.

In practical terms, they are the local halves
of shovels or federation links, whereas the
remote parts use regular ("network") connections
that open new TCP connections, like any
AMQP 0-9-1 client would.

Because direct connections are not necessarily
backed by a network connection (imagine a
shovel that uses a queue with a leader
hosted on the very same node), they may or may
not have certain fields that responses::Connection
was expecting.

This solution is more straightforward and less
disruptive compared to using an enum that
represents both connection types, and trying to
implement Tabled for it.

Closes #68.